Debug

+++ app/src/main/java/com/example/attendify/ClassItem.java
@@ -2,12 +2,12 @@ package com.example.attendify;
 
 public class ClassItem {
     private long cid;
-
     public ClassItem(long cid, String className, String subjectName) {
         this.cid = cid;
         this.className = className;
         this.subjectName = subjectName;
     }
+
 
     private String className;
 
@@ -18,27 +18,26 @@ public class ClassItem {
     public void setClassName(String className) {
         this.className = className;
     }
-
-    public String getSubjectName() {
+    public String getSubjectName(){
         return subjectName;
     }
-
-    public void setSubjectName(String subjectName) {
+    public void setSubjectName(String subjectName){
         this.subjectName = subjectName;
     }
-
     private String subjectName;
 
-    public ClassItem(String className, String subjectName) {
+    public ClassItem(String className, String subjectName){
         this.className = className;
         this.subjectName = subjectName;
     }
+
+
 
     public long getCid() {
         return cid;
     }
 
-    public void setCid(int cid) {
+    public void setCid(long cid) {
         this.cid = cid;
     }
 }
